:truck: Chore - Move test folder
