+++ server/test/integration/helper.js
@@ -1,7 +1,7 @@
 /* eslint import/no-extraneous-dependencies: off */
 import chai from 'chai';
 import chaiHttp from 'chai-http';
-import app from '../../app';
+import app from '../../src/app';
 
 chai.use(chaiHttp);
 export const {
