+++ server/test/unit/controllers/controller.js
@@ -2,7 +2,7 @@
 import {
   expect
 } from 'chai';
-import Controller from '../../../controllers/controller';
+import Controller from '../../../src/controllers/controller';
 
 describe('Controller', () => {
